pyproject.toml
[tool.poetry]
name = "annotation-backend"
version = "0.1.0"
description = ""
authors = ["OpenNyAI Team <opennyai@googlegroups.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pydantic = "1.10.13"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
fastapi = "0.100.0"
passlib = "^1.7.4"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
greenlet = "^3.0.3"
bcrypt = "^4.1.3"
sqlalchemy = "^2.0.30"
python-jose = "^3.3.0"
sendgrid = "^6.11.0"
pytz = "^2024.1"
pymupdf = "^1.24.5"
docx2txt = "^0.8"
psycopg = {extras = ["binary", "pool"], version = "^3.1.19"}
langchain-community = "^0.2.5"
langchain = "^0.2.5"
langchain-openai = "^0.1.9"
langchain-postgres = "^0.0.9"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.4"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.3.0"
mypy = "^1.3.0"
poethepoet = "^0.20.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"
gunicorn = "^20.1.0"
faker = "^18.11.1"
pytest-celery = "^0.0.0"
oauth2client = "^4.1.3"
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks.lint]
shell = """
black app tests
flake8 app tests
mypy app tests
"""
interpreter = "bash"
help = "format, lint, typecheck"


[tool.poe.tasks.test]
cmd = "python -m pytest -vv -o log_cli=1 -o log_cli_level=INFO -W 'ignore::DeprecationWarning' $FILE"
args = [{name="FILE", default="tests", positional=true}]
help = "run tests using pytest"
envfile = "tests/.env"

[tool.poe.tasks.start]
shell = """
uvicorn --port 8080 --reload app:app
"""
help = "start the server for local development"


[tool.poe.tasks.build]
shell = """
docker-compose build api
"""
help = "build docker container"
